requirements.txt
gradio==4.0.0
torch
transformers
sentence-transformers
faiss-cpu
pymupdf
numpy
